requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-dotenv
langchain
langchain-community
langchain-openai
langchain-experimental
openai
pandas
matplotlib
plotly
seaborn
numpy
requests
